backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1
sqlalchemy==2.0.23
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.1.0.post1
supabase-py==2.0.3 